Ported tensor_to_string_list to tensor_utils
